feat: change models

+++ project/app/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class UserProfile(models.Model):
     username = models.CharField('username',max_length=100)
     email = models.EmailField('email',max_length=100)
-    phone= models.IntegerField('phone')
+    phone= models.BigIntegerField('phone')
 
     def __str__(self):
         return f"Name: {self.username}, Email: {self.email}, Phone: {self.phone}"
